+++ src/Multiplayer/Cheat/MpCheatApplier.cs
@@ -59,7 +59,9 @@ public static class MpCheatApplier {
         CheatsActive && MpCheatState.Config.GlobalMap.UnknownMapAlwaysTreasure;
 
     public static bool FreeTravelFromDevRoomMap =>
-        CheatsActive && MpCheatState.Config.GlobalMap.FreeTravelFromDevRoomMap;
+        CheatsActive && (
+            MpCheatState.Config.GlobalMap.FreeTravelFromDevRoomMap
+            || DevModeState.MapCheats.FreeTravelFromDevRoomMap);
 
     public static bool FrameCheatsAllowed =>
         CheatsActive && !MpCheatSession.InMultiplayerRun;